Finished version
The b-tree index with attribute a as the key value is completed, and the multi-channel merge sorting is completed. Other requirements of the experiment are also met. Finally, it runs successfully under the data test. Except the simple algorithm running speed of multi-channel merge sorting, all of them are acceptable